fix(statistics): cast operands to double to prevent potential integer overflow

+++ src/ExperimentFramework.Science/Statistics/MannWhitneyUTest.cs
@@ -88,8 +88,8 @@ public sealed class MannWhitneyUTest : IStatisticalTest
         if (tieCorrection > 0)
         {
             var n = n1 + n2;
-            // Cast to double before multiplication to avoid potential integer overflow
-            stdU = Math.Sqrt((double)n1 * n2 / ((double)n * (n - 1)) *
+            // Cast both operands to double before multiplication to avoid potential integer overflow
+            stdU = Math.Sqrt((double)n1 * (double)n2 / ((double)n * (double)(n - 1)) *
                 ((Math.Pow(n, 3) - n) / 12 - tieCorrection));
         }
